// client/model/aciRuntime.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EndpointGroup } from './endpointGroup';
import { Vlan } from './vlan';


export interface AciRuntime { 
    readonly id?: string;
    macAddress: string;
    addr?: string;
    name: string;
    contName?: string;
    lcC?: Array<string>;
    readonly vlan?: Vlan;
    vlanId?: string;
    readonly endpointGroup?: EndpointGroup;
    endpointGroupId?: string;
    runtimeDataLastRefreshed?: string;
}